+++ app/controllers/questions_controller.rb
@@ -1,10 +1,13 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { question.answers.new }
 
   def create
+    question.author = current_user
     if question.save
-      redirect_to questions_path
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -19,8 +22,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question successfully deleted.'
+    else
+      redirect_to question
+    end
   end
 
   private
